requirements.txt
pynput
screeninfo
Pillow
pywin32
psutil
pystray
